Push native-width types through leaf APIs
Remove compatibility shims where syscall code cast usize back to u64
for lower-level APIs. Now native types flow end-to-end:

- fill_linux_stat: buf u64 → usize (pointer to user buffer)
- is_serial_fd: fd u64 → usize
- execargs::set_from_user: argv_ptr/envp_ptr u64 → usize
- ArchSpecificOps trait: nr/a0/a1 u64 → usize, return i64 → isize
- MMAP_BASE, PROGRAM_BRK: u64 → usize (virtual addresses)
- VFORK_PARENT_MMAP_BASE/BRK: u64 → usize
- map_user_pages: start_va/end_va u64 → usize

Remaining u32 casts (flags, whence, mode) are correct — these are
genuinely 32-bit per the POSIX/Linux ABI on all architectures.

+++ crates/arch/src/lib.rs
@@ -134,7 +134,7 @@ pub trait ArchInfo {
 /// Architecture-specific syscalls (e.g., arch_prctl on x86_64).
 /// Returns Some(result) if handled, None if not recognized.
 pub trait ArchSpecificOps {
-    fn arch_syscall(nr: u64, a0: u64, a1: u64) -> Option<i64>;
+    fn arch_syscall(nr: usize, a0: usize, a1: usize) -> Option<isize>;
 }
 
 /// Vfork/exec context: save and restore arch-specific process state.
